% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_pages.R
\name{bind_pages}
\alias{bind_pages}
\title{Bind multiple pages by row}
\usage{
bind_pages(pages)
}
\arguments{
\item{pages}{List of data frames representing Megamation API pages.}
}
\description{
Bind multiple pages by row
}
