% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{body_parse}
\alias{body_parse}
\title{Parse body from Megamation API response}
\usage{
body_parse(resp_body)
}
\arguments{
\item{resp_body}{An API response body.}
}
\value{
A list.
}
\description{
\code{body_parse()} converts the raw bytes of an API response to an R list object.
After converting bytes to characters, encoding is done to resolve
a UTF-8 issue from Megamation's side.
}
