% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_bool}
\alias{check_bool}
\title{Is boolean (length-1 logical)}
\usage{
check_bool(x, arg = rlang::caller_arg(x), call = rlang::caller_env())
}
\description{
Is boolean (length-1 logical)
}
\keyword{internal}
