% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_date}
\alias{check_date}
\title{Title Check date-like inputs and convert them}
\usage{
check_date(x, arg = caller_arg(x), call = caller_env())
}
\arguments{
\item{arg}{Date, POSIX(c/l)t date/time, or parse-able string in
YYYY(/-)MM(/-)DD format.
Intended to be converted to MM-DD-YYYY format.}
}
\value{
single string date/time formatted in ISO8601
}
\description{
Title Check date-like inputs and convert them
}
\keyword{internal}
