% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_params}
\alias{check_params}
\title{Are httr2 parameters well-specified}
\usage{
check_params(params, call = rlang::caller_env())
}
\arguments{
\item{params}{A list of parameter name-value pairs.}
}
\value{
A list of parameter name-value pairs.
}
\description{
\code{format_params()} formats supplied name-value pairs toward
creating a valid Megamation URL.
}
\keyword{internal}
