% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_url}
\alias{check_url}
\title{Checking and re-formatting URL}
\usage{
check_url(x, arg = rlang::caller_arg(x), call = rlang::caller_env())
}
\description{
Checking and re-formatting URL
}
\keyword{internal}
