% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{date_as_between_string}
\alias{date_as_between_string}
\title{Format sequence of dates}
\usage{
date_as_between_string(.min, .max)
}
\arguments{
\item{.min}{A minimum Date.}

\item{.max}{A maximum Date.}
}
\value{
A single string formatted as "<>MM-DD-YYYY,MM-DD-YYYY" where the
"<>" modifiers means "in-between".
}
\description{
\code{date_as_between_string()} formats a sequence of dates to a single string
that represents the sequence. The purpose is to simplify the request URL from
including \code{DATE=X&DATE=Y&DATE=Z} to \verb{DATE = <>X,Z} where X<Y<Z.
}
\examples{
from <- as.Date("2023-09-20")
to <- as.Date("2023-10-20")
date_as_between_string(from, to)
}
\keyword{internal}
