% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_data.R
\name{find_data}
\alias{find_data}
\title{Find data in the JSON list returned by Megamation's API}
\usage{
find_data(json_list)
}
\arguments{
\item{json_list}{A JSON list.}
}
\description{
Find data in the JSON list returned by Megamation's API
}
