% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{format_date}
\alias{format_date}
\title{Format date values}
\usage{
format_date(date)
}
\arguments{
\item{date}{A vector of type Date.}
}
\value{
A vector of strings.
}
\description{
\code{format_date()} formats given date values to be compliant with the API (
MM-DD-YYYY).
}
\examples{
# Single date
format_date(as.Date("2023-09-20"))

# Sequence of dates
from <- as.Date("2023-09-20")
to <- as.Date("2023-10-20")
date <- seq(from, to, by = "day")
format_date(date)
}
\keyword{internal}
