% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env.R
\name{get_env_key}
\alias{get_env_key}
\title{Get \code{MEGAMATION_KEY} env var}
\usage{
get_env_key()
}
\value{
The string value of the MEGAMATION_KEY environment variable
or an error if none exists.
}
\description{
Get \code{MEGAMATION_KEY} env var
}
\keyword{internal}
