% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env.R
\name{get_env_url}
\alias{get_env_url}
\title{Get \code{MEGAMATION_URL} env var}
\usage{
get_env_url()
}
\value{
The string value of the MEGAMATION_URL environment variable
or an error if none exists.
}
\description{
Get \code{MEGAMATION_URL} env var
}
\keyword{internal}
