% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env.R
\name{get_env_user}
\alias{get_env_user}
\title{Get \code{MEGAMATION_USER} env var}
\usage{
get_env_user()
}
\value{
The string value of the MEGAMATION_USER environment variable
or an error if none exists.
}
\description{
Get \code{MEGAMATION_USER} env var
}
\keyword{internal}
