% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/megamation-package.R
\docType{package}
\name{megamation-package}
\alias{megamation}
\alias{megamation-package}
\title{megamation: Functions for downloading Megamation data directly in R}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/asadow/megamation}
  \item \url{https://asadow.github.io/megamation/}
  \item Report bugs at \url{https://github.com/asadow/megamation/issues}
}

}
\author{
\strong{Maintainer}: Adam Sadowski \email{asadowsk@uoguelph.ca} (\href{https://orcid.org/JANE'S-ORCID-ID}{ORCID})

}
\keyword{internal}
