% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{mm_args_for_gets}
\alias{mm_args_for_gets}
\title{Get arguments for GET request(s)}
\usage{
mm_args_for_gets(endpoint, ...)
}
\arguments{
\item{endpoint}{The API endpoint. For example,
\code{"timecard"} for employee transactions, and \code{"workorder"}
for work orders. All endpoints are listed at
https://apidocs.megamation.com/.}

\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> Name-value pairs to filter the
request. The name should be the lower-case name of a field that is
filter-enabled (in Megamation's words, a criteria). These arguments are
processed with \code{\link[rlang:defusing-advanced]{rlang::quos()}} and support unquote via \code{\link{!!}} and
unquote-splice via \code{\link{!!!}}.}
}
\value{
A list of dataframes where each column is an argument
to \code{\link[=mm_req]{mm_req()}}.
}
\description{
For fetching a specified endpoint with parameters, \code{mm_args_for_gets()} will
return the argument dataframes needed for any GET requests.
}
\examples{
# Return arguments needed for timecard GET for Jan. 2022 to 2023:

# Create a Date-type vector
date <- seq.Date(
  as.Date("2022-01-01"),
  as.Date("2023-01-31"),
  by = "day"
)

mm_args_for_gets("timecard", date = date)
}
