% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authorize.R
\name{mm_authorize}
\alias{mm_authorize}
\title{Install Megamation credentials}
\usage{
mm_authorize(key, url)
}
\arguments{
\item{key}{Single string of the Megamation API key.}

\item{url}{Single string of the Megamation API base URL.}
}
\value{
Writes \code{MEGAMATION_KEY} and \code{MEGAMATION_URL} environment variables.
}
\description{
\code{mm_authorize()} installs your credentials so that \code{megamation} functions
that need them, like \code{\link[=mm_get]{mm_get()}}, can use them automatically.

Specifically, it places your Megamation API key and base URL
in your computer's \code{megamation} R package configuration files. Then,
whenever the package is loaded, your credentials are available as environment
variables.
}
\examples{
\dontrun{
mm_authorize(
  key = "<YOUR-MEGAMATION_KEY>",
  url = "<YOUR-MEGAMATION_URL>"
)
}
}
