% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mm_bind_then_tbl.R
\name{mm_bind_then_tbl}
\alias{mm_bind_then_tbl}
\title{Bind multiple Megamation API pages by row before converting to a tibble}
\usage{
mm_bind_then_tbl(pages)
}
\arguments{
\item{pages}{List of data frames representing Megamation API pages.}
}
\description{
Bind multiple Megamation API pages by row before converting to a tibble
}
