% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mm_error_body.R
\name{mm_error_body}
\alias{mm_error_body}
\title{Extract error body from a Megamation response}
\usage{
mm_error_body(resp)
}
\arguments{
\item{resp}{An API response.}
}
\description{
Extract error body from a Megamation response
}
