% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_col_info.R
\name{mm_get_appendix}
\alias{mm_get_appendix}
\title{Get endpoint appendix (criteria, schema, labels)}
\usage{
mm_get_appendix(endpoint, appendix)
}
\arguments{
\item{endpoint}{The API endpoint. For example,
\code{"timecard"} for employee transactions, and \code{"workorder"}
for work orders. All endpoints are listed at
https://apidocs.megamation.com/.}

\item{appendix}{\code{"criteria"}, \code{"labels"}, or \code{"schema"}.}
}
\value{
A data frame of class \code{\link[tibble:tbl_df-class]{tbl_df}}
containing the endpoint's appendix.
}
\description{
\code{mm_get_appendix()} returns the user-supplied appendix.
}
\keyword{internal}
