% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform.R
\name{mm_get_names}
\alias{mm_get_names}
\title{Get column names and which are currently filters}
\usage{
mm_get_names(endpoint)
}
\arguments{
\item{endpoint}{The API endpoint. For example,
\code{"timecard"} for employee transactions, and \code{"workorder"}
for work orders. All endpoints are listed at
https://apidocs.megamation.com/.}
}
\value{
A data frame of class \code{\link[tibble:tbl_df-class]{tbl_df}}
containing the requested information.
}
\description{
\code{mm_get_names()} performs and combines two separate GET requests
on an endpoint: one
for Schema and one for Criteria. Schema returns the names and (database)
types of all available columns. Criteria returns the names of columns for
which filtering is enabled.
}
\examples{
\dontrun{
mm_get_names("status")
}
}
