% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authorize.R
\name{mm_has_creds}
\alias{mm_has_creds}
\title{Are there credentials on hand?}
\usage{
mm_has_creds()
}
\value{
\code{TRUE} if user has \code{MEGAMATION_KEY} and \code{MEGAMATION_URL} environment
variables.
}
\description{
Are there credentials on hand?
}
\concept{low-level API functions}
\keyword{internal}
