% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mm_keep_df.R
\name{mm_keep_df}
\alias{mm_keep_df}
\title{Extract data in the JSON list returned by Megamation's API}
\usage{
mm_keep_df(list_from_json)
}
\arguments{
\item{list_from_json}{A list returned by \code{jsonlite::fromJSON()}.}
}
\description{
Extract data in the JSON list returned by Megamation's API
}
\keyword{internal}
