% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env.R
\name{mm_key}
\alias{mm_key}
\title{Get Megamation key}
\usage{
mm_key()
}
\value{
The string value of the \code{MEGAMATION_KEY} environment variable or an
error if none exists.
}
\description{
Get Megamation key
}
\keyword{internal}
