% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names.R
\name{mm_names}
\alias{mm_names}
\title{Retrieve column names and details}
\usage{
mm_names(endpoint)
}
\arguments{
\item{endpoint}{The API endpoint. For example, \code{"timecard"} for employee
transactions, and \code{"workorder"} for work orders. All endpoints are listed
at \url{https://apidocs.megamation.com/}.}
}
\value{
A data frame of class \code{\link[tibble:tbl_df-class]{tbl_df}} containing
the endpoint's appendix.
}
\description{
\code{mm_names()} returns column names, filter status, description,
and types through GET requests of the criteria and schema appendices.
}
\examples{
\dontshow{if (megamation:::has_creds()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
mm_names("status")
\dontshow{\}) # examplesIf}
}
