% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paginate.R
\name{mm_next_req}
\alias{mm_next_req}
\title{Return next request (if existing) from a Megamation API response}
\usage{
mm_next_req(resp, req)
}
\arguments{
\item{resp}{An API response.}

\item{req}{An API request generated by \code{\link[=mm_request]{mm_request()}}.}
}
\value{
An object of class \code{httr2_request} or \code{NULL}.
}
\description{
\code{mm_next_req()} returns the next request based on the next_url given
by the parsed response.
}
\examples{
\dontrun{
mm_request("workorder") |>
  httr2::req_perform_iterative(
    next_req = function(resp, req) mm_next_req(resp, req)
    )
}
}
