% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pagebind.R
\name{mm_pagebind}
\alias{mm_pagebind}
\title{Bind multiple Megamation API pages by row before converting to a tibble}
\usage{
mm_pagebind(pages)
}
\arguments{
\item{pages}{List of data frames representing Megamation API pages.}
}
\value{
A data frame of class \code{\link[tibble:tbl_df-class]{tbl_df}} representing
the bound pages.
}
\description{
\code{mm_pagebind()} is needed as pages can have same-named fields with different
types. This is because some field(s) of a given page may or may not contain
vectors  of values in one of its rows. \code{mm_pagebind()} takes care of this
possibility by treating each page as a matrix before binding and unnesting
their combinations.
}
\examples{
page_1 <- mtcars |>
  dplyr::mutate(
    mpg = as.list(mpg),
  )
page_2 <- page_1 |>
  dplyr::mutate(
    cyl = list(cyl)
  )
pages <- list(page_1, page_2)
# mpg unnests but not cyl
megamation:::mm_pagebind(pages)
}
\keyword{internal}
