% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{mm_parse}
\alias{mm_parse}
\title{Parse the Megamation API response}
\usage{
mm_parse(resp_body)
}
\arguments{
\item{resp_body}{A Megamation API response body.}
}
\description{
The body of the response contains raw bytes.
After converting these bytes to a string, encoding is needed due to
a UTF-8 issue from Megamation's side.
}
