% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-parse.R
\name{mm_parsed_extract}
\alias{mm_parsed_extract}
\alias{extract_data}
\alias{extract_criteria_or_labels}
\alias{extract_schema}
\title{Create tibble from parsed response}
\usage{
mm_parsed_extract(parsed, .get = "data")

extract_data(parsed)

extract_criteria_or_labels(parsed)

extract_schema(parsed)
}
\arguments{
\item{parsed}{Parsed response body.}

\item{.get}{Whether the GET request is for the endpoint's \code{"data"},
\code{"criteria"}, \code{"labels"}, or \code{"schema"}.}
}
\value{
A data frame containing the endpoint data.
}
\description{
\code{mm_parsed_extract()} extracts a data frame from the parsed response body.
}
\examples{
# Returns NULL from empty list
resp <- httr2::response_json()
resp |>
  megamation:::mm_resp_parse() |>
  megamation:::mm_parsed_extract()
}
\keyword{internal}
