% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mm_pluck_data.R
\name{mm_pluck_data}
\alias{mm_pluck_data}
\title{Extract data in the JSON list returned by Megamation's API}
\usage{
mm_pluck_data(list_from_json)
}
\arguments{
\item{list_from_json}{A list returned by \code{jsonlite::fromJSON()}.}
}
\description{
Extract data in the JSON list returned by Megamation's API
}
\keyword{internal}
