% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mm_pull.R
\name{mm_pull}
\alias{mm_pull}
\title{Define and perform a Megamation API request}
\usage{
mm_pull(resource, ..., url = get_url(), user = get_user(), key = get_key())
}
\arguments{
\item{resource}{The resource endpoint. For example,
"timecard" for employee transactions, and
"workorder" for work orders. All endpoints are listed here:
"https://apidocs.megamation.com/".}

\item{...}{API parameters to filter the request.}

\item{user}{The API key user.}

\item{api_key}{The API key.}
}
\description{
Define and perform a Megamation API request
}
