% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify.R
\name{mm_req_append}
\alias{mm_req_append}
\title{Append a GET request}
\usage{
mm_req_append(req, x)
}
\arguments{
\item{req}{An API request generated by \code{\link[=mm_request]{mm_request()}}.}

\item{x}{\code{"criteria"}, \code{"labels"}, or \code{"schema"}.}
}
\value{
An object of class \code{httr2_request}.
}
\description{
Append a GET request
}
\examples{
\dontshow{if (httr2::secret_has_key("HTTR2_KEY_MEGAMATION")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
mm_request("workorder") |> mm_req_append("criteria")
\dontshow{\}) # examplesIf}
}
