% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paginate.R
\name{mm_req_paginate}
\alias{mm_req_paginate}
\title{Add pagination to a Megamation API request}
\usage{
mm_req_paginate(req)
}
\arguments{
\item{req}{An API request generated by \code{\link[=mm_req]{mm_req()}} or \code{\link[httr2:req]{httr2::req()}}.}
}
\description{
Add pagination to a Megamation API request
}
