% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paginate.R
\name{mm_req_paginate_custom}
\alias{mm_req_paginate_custom}
\title{Add pagination to a Megamation API request}
\usage{
mm_req_paginate_custom(req)
}
\arguments{
\item{req}{An API request generated by \code{\link[=mm_request]{mm_request()}}.}
}
\value{
An object of class \code{httr2_request}.
}
\description{
Add pagination to a Megamation API request
}
