% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify.R
\name{mm_req_params}
\alias{mm_req_params}
\title{Modify request URL with filtering components}
\usage{
mm_req_params(req, ..., allfields = TRUE)
}
\arguments{
\item{req}{An API request generated by \code{\link[=mm_req]{mm_req()}}.}

\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> Name-value pairs to filter the
request. The name should be the lower-case name of a field that is
filter-enabled (in Megamation's words, a criteria). These arguments are
processed with \code{\link[rlang:defusing-advanced]{rlang::quos()}} and support unquote via \code{\link{!!}} and
unquote-splice via \code{\link{!!!}}.}

\item{allfields}{If \code{TRUE}, return all fields currently available for
the endpoint.}
}
\value{
An object of class \code{httr2_request} with a pagination policy.
}
\description{
\code{mm_req_params()} adds filters to the request. By default, it adds the query
for all (currently available) fields.
}
\keyword{internal}
