% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform.R
\name{mm_req_perform}
\alias{mm_req_perform}
\title{Perform a Megamation API request}
\usage{
mm_req_perform(req)
}
\arguments{
\item{req}{An API request generated by \code{\link[=mm_request]{mm_request()}}.}
}
\value{
An \link[httr2:response]{httr2::response} or a list of such responses where the request
is paginated.
}
\description{
After creating a request with \code{\link[=mm_request]{mm_request()}},
call \code{mm_req_perform()} to perform it and fetch
the results back to R.
}
