% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paginate.R
\name{mm_req_perform_paginate}
\alias{mm_req_perform_paginate}
\title{Perform a paginated request}
\usage{
mm_req_perform_paginate(req, max_pages = NULL)
}
\arguments{
\item{req}{An API request generated by mm_req() with pagination from
mm_req_paginate().}

\item{max_pages}{Max number of pages to request.}
}
\description{
Perform a paginated request
}
