% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paginate.R
\name{mm_req_perform_paginate_custom}
\alias{mm_req_perform_paginate_custom}
\title{Perform a paginated request}
\usage{
mm_req_perform_paginate_custom(req, max_pages = NULL)
}
\arguments{
\item{req}{An API request generated by \code{\link[=mm_req]{mm_req()}} with pagination from
\code{\link[=mm_req_paginate]{mm_req_paginate()}}.}

\item{max_pages}{Max number of pages to request.}
}
\value{
A list of data frames representing the pages of responses.
}
\description{
Perform a paginated request
}
