% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mm_req_wo.R
\name{mm_req_workorder}
\alias{mm_req_workorder}
\title{Extract Megamation workorders}
\usage{
mm_req_workorder(url, status = NULL, type = NULL, date)
}
\arguments{
\item{url}{A URL starting with "https://api.megamation.com/".}

\item{status}{Work order status. For example, S.}

\item{type}{Work order type.}

\item{date}{Work order date.}
}
\description{
Extract Megamation workorders
}
