% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{mm_request}
\alias{mm_request}
\title{Create a Megamation API request}
\usage{
mm_request(endpoint, .url = get_env_url(), .key = get_env_key())
}
\arguments{
\item{endpoint}{The API endpoint. For example,
\code{"timecard"} for employee transactions, and \code{"workorder"}
for work orders. All endpoints are listed at
https://apidocs.megamation.com/.}

\item{.url}{API base URL for request.}

\item{.key}{API key for request.}
}
\value{
An object of class \code{httr2_request}.
}
\description{
\code{mm_request()} creates a request using \code{\link[httr2:request]{httr2::request()}} and
does the following:
\itemize{
\item Inserts the base URL using the environment variable \code{MEGAMATION_URL} from
your \code{.Renviron}. Your key and base URL can be set using \code{\link[=mm_set_creds]{mm_set_creds()}}.
\item Appends the URL with the endpoint defined by parameter \code{endpoint}.
\item Sets the user-agent as the GitHub \code{megamation} package.
\item Authenticates the request with HTTP basic authentication using
environment variables \code{MEGAMATION_KEY} and \code{MEGAMATION_URL}
from your \code{.Renviron}.
\item Handles HTTP errors so useful information from the response is extracted
(e.g. "No response body").
\item Adds caching of responses if available. See \code{\link[httr2:req_cache]{httr2::req_cache()}}.
}
}
\examples{
\dontshow{if (httr2::secret_has_key("HTTR2_KEY_MEGAMATION")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
mm_request("timecard")
mm_request("trade")
\dontshow{\}) # examplesIf}
}
