% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{mm_resp_body}
\alias{mm_resp_body}
\title{Extract and parse body from Megamation API response}
\usage{
mm_resp_body(resp)
}
\arguments{
\item{resp}{A Megamation API response.}
}
\description{
The body of the response contains raw bytes.
After converting these bytes to a string, encoding is done to resolve
a UTF-8 issue from Megamation's side.
}
