% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{mm_resp_data}
\alias{mm_resp_data}
\title{Extract data from Megamation response}
\usage{
mm_resp_data(resp)
}
\arguments{
\item{resp}{A Megamation API response.}
}
\description{
Extract data from Megamation response
}
