% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mm_resp_process.R
\name{mm_resp_process}
\alias{mm_resp_process}
\title{Process a Megamation API response}
\usage{
mm_resp_process(resp)
}
\arguments{
\item{resp}{A Meagmation API response.}
}
\description{
Process a Megamation API response
}
