% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env.R
\name{mm_url}
\alias{mm_url}
\title{Get Megamaton URL}
\usage{
mm_url()
}
\value{
The string value of the \code{MEGAMATION_URL} environment variable
or an error if none exists.
}
\description{
Get Megamaton URL
}
\keyword{internal}
