% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parsed_extract}
\alias{parsed_extract}
\alias{extract_data}
\alias{extract_criteria}
\alias{extract_schema}
\title{Create tibble from parsed response}
\usage{
parsed_extract(parsed, .get = c("data", "criteria", "labels", "schema"))

extract_data(parsed)

extract_criteria(parsed)

extract_schema(parsed)
}
\arguments{
\item{parsed}{Parsed response body.}
}
\value{
A data frame containing the endpoint data.
}
\description{
\code{parsed_extract()} extracts a data frame from the parsed
response body.
}
