% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mm_req_wo.R
\name{req_mm_wo}
\alias{req_mm_wo}
\title{Perform a Megamation API workorder request}
\usage{
req_mm_wo(url, status = NULL, type = NULL, date)
}
\arguments{
\item{url}{A URL starting with "https://api.megamation.com/".}

\item{status}{Work order status. For example, S.}

\item{type}{Work order type.}

\item{date}{Work order date.}
}
\description{
Perform a Megamation API workorder request
}
