% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opts.R
\name{req_opts}
\alias{req_opts}
\title{Specify request options}
\usage{
req_opts(.url = get_env_url(), .key = get_env_key())
}
\arguments{
\item{.url}{API base URL for request.}

\item{.key}{API key for request.}
}
\description{
Specify request options
}
\examples{
req_opts(.url = "base-url")
}
