% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{resp_body_parse}
\alias{resp_body_parse}
\title{Extract and parse body from Megamation API response}
\usage{
resp_body_parse(resp)
}
\arguments{
\item{resp}{An API response.}
}
\value{
A list.
}
\description{
The body of the response contains raw bytes.
After converting these bytes to a string, encoding is done to resolve
a UTF-8 issue from Megamation's side.
}
\details{
\code{resp_body_parse()} extracts the raw bytes from an API response and parses
it, returning an R list object.
}
