% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_api_key.R
\name{set_api_key}
\alias{set_api_key}
\title{Set MEGATION_KEY as an environment variable}
\usage{
set_api_key(key = NULL)
}
\arguments{
\item{key}{The API key.}
}
\description{
Set MEGATION_KEY as an environment variable
}
