% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env.R
\name{testing_key}
\alias{testing_key}
\title{Get personal key}
\usage{
testing_key()
}
\value{
A string of a decrypted key.
}
\description{
testing_key() uses the HTTR2_KEY_MEGAMATION environment
variable to decrypt a secret.
}
\keyword{internal}
