% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{to_tbl_criteria}
\alias{to_tbl_criteria}
\title{Create tibble from parsed response to Get Criteria or Get Labels}
\usage{
to_tbl_criteria(parsed)
}
\arguments{
\item{parsed}{Parsed response body.}
}
\value{
A \link{tbl_df} with columns \code{list_name} and \code{value}
representing the Criteria or Labels response list.
}
\description{
Create tibble from parsed response to Get Criteria or Get Labels
}
