% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{to_tbl_data}
\alias{to_tbl_data}
\title{Create tibble from parsed response}
\usage{
to_tbl_data(parsed)
}
\arguments{
\item{parsed}{Parsed response body.}
}
\value{
A \link{tbl_df} of the \link{data.frame} object contained in the
parsed response. Columns are unaltered from the endpoint's data.
}
\description{
Create tibble from parsed response
}
