% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{to_tbl_schema}
\alias{to_tbl_schema}
\title{Create tibble from parsed response to Get Schema}
\usage{
to_tbl_schema(parsed)
}
\arguments{
\item{parsed}{Parsed response body.}
}
\value{
A \link{tbl_df} with columns \code{list_name} and \code{value}
representing the Schema response list. \code{value} is
\code{value} is of type \code{list} because of the lists contained in the values
indexed by the \code{required} and \code{properties} list names.
}
\description{
Create tibble from parsed response to Get Schema
}
