% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_date}
\alias{check_date}
\title{Check date-like inputs}
\usage{
check_date(x, arg = rlang::caller_arg(x), call = rlang::caller_env())
}
\arguments{
\item{arg}{Date vector.}
}
\value{
single string date/time formatted in ISO8601
}
\description{
Check date-like inputs
}
\keyword{internal}
