% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_string}
\alias{check_string}
\title{Is string (length-1 character)}
\usage{
check_string(
  x,
  arg = rlang::caller_arg(x),
  call = rlang::caller_env(),
  optional = FALSE
)
}
\description{
Is string (length-1 character)
}
\keyword{internal}
