% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-params.R
\name{format_params}
\alias{format_params}
\title{Format a list of parameters}
\usage{
format_params(endpoint, ...)
}
\arguments{
\item{endpoint}{The API endpoint. For example,
\code{"timecard"} for employee transactions, and \code{"workorder"}
for work orders. All endpoints are listed at
https://apidocs.megamation.com/.}

\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> Name-value pairs to filter the
request. The name should be the lower-case name of a field that is
filter-enabled (in Megamation's words, a criteria). These arguments are
processed with \code{\link[rlang:defusing-advanced]{rlang::quos()}} and support unquote via \code{\link{!!}} and
unquote-splice via \code{\link{!!!}}.}
}
\value{
A list of parameter name-value pairs.
}
\description{
\code{format_params()} formats supplied name-value pairs toward
creating a valid and readable Megamation URL.
}
\examples{
from <- as.Date("2023-09-20")
to <- as.Date("2024-10-20")
date <- seq(from, to, by = "day")
trade <- c("[]PCO", "[]DM")
megamation:::format_params("workorder", date = date, trade = trade)
}
\keyword{internal}
