% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{format_params}
\alias{format_params}
\title{Format a list of parameters}
\usage{
format_params(...)
}
\arguments{
\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> Name-value pairs to filter the request.
The name should be the lower-case name of a
field that is filter-enabled
(in Megamation's words, a criteria).}
}
\value{
A list of parameter name-value pairs.
}
\description{
\code{format_params()} formats supplied name-value pairs toward
creating a valid Megamation URL.
}
\examples{
from <- as.Date("2023-09-20")
to <- as.Date("2023-10-20")
date <- seq(from, to, by = "day")
trade <- c("[]PCO", "[]DM")
format_params(date = date, trade = trade)
}
