% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paginate.R
\name{is_paginated}
\alias{is_paginated}
\title{Detect whether a request is paginated}
\usage{
is_paginated(req)
}
\arguments{
\item{req}{An API request.}
}
\value{
\code{TRUE} or \code{FALSE}.
}
\description{
\code{is_paginated()} checks whether an \code{httr2_request} has a paginate policy.
}
\examples{
mm_req("workorder") |> is_paginated()
mm_req("workorder") |> mm_req_paginate() |> is_paginated()
}
