% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_paginated.R
\name{is_paginated}
\alias{is_paginated}
\title{Detect whether a request is paginated}
\usage{
is_paginated(x)
}
\arguments{
\item{x}{An API request.}
}
\description{
Detect whether a request is paginated
}
