% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-check.R
\name{is_paginated}
\alias{is_paginated}
\title{Detect whether a request is paginated
\code{is_paginated()} checks whether an \code{httr2_request} has a paginate policy.}
\usage{
is_paginated(req)
}
\arguments{
\item{req}{httr2 request.}
}
\value{
\code{TRUE} or \code{FALSE}.
}
\description{
Detect whether a request is paginated
\code{is_paginated()} checks whether an \code{httr2_request} has a paginate policy.
}
