% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paginate.R
\name{is_paginated}
\alias{is_paginated}
\title{Detect whether a request is paginated}
\usage{
is_paginated(req)
}
\arguments{
\item{req}{An API request generated by \code{\link[=mm_req]{mm_req()}}.}
}
\value{
\code{TRUE} or \code{FALSE}.
}
\description{
\code{is_paginated()} checks whether an \code{httr2_request} has a paginate policy.
}
\examples{
\dontshow{if (httr2::secret_has_key("HTTR2_KEY")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
mm_req("workorder") |> is_paginated()
mm_req("workorder") |> mm_req_paginate() |> is_paginated()
\dontshow{\}) # examplesIf}
}
\keyword{internal}
