% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/megamation-package.R
\docType{package}
\name{megamation-package}
\alias{megamation}
\alias{megamation-package}
\title{megamation: Access `DirectLine` Data Using the `Megamation API`}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Tools for accessing and handling `DirectLine` data using the `Megamation API` \url{https://apidocs.megamation.com/}. The API allows users to interact with data programmatically, instead of via the `DirectLine` software. This package is community-maintained and is not officially supported by Megamation \url{https://megamation.com/about-megamation/}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/asadow/megamation}
  \item \url{https://asadow.github.io/megamation/}
  \item Report bugs at \url{https://github.com/asadow/megamation/issues}
}

}
\author{
\strong{Maintainer}: Adam Sadowski \email{asebsadowski@gmail.com} (\href{https://orcid.org/0000-0002-8023-9227}{ORCID})

Other contributors:
\itemize{
  \item University of Guelph [copyright holder, funder]
}

}
\keyword{internal}
