% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paginate.R
\name{mm_bind_then_tbl}
\alias{mm_bind_then_tbl}
\title{Bind multiple Megamation API pages by row before converting to a tibble}
\usage{
mm_bind_then_tbl(pages)
}
\arguments{
\item{pages}{List of data frames representing Megamation API pages.}
}
\value{
A data frame of class \link{tbl} representing the bound pages.
}
\description{
\code{mm_bind_then_tbl()} is needed as pages can have same-named columns with
different types. This is because some column(s) of a given page
may or may not contain vectors  of values in one of its row.
\code{mm_bind_then_tbl()} takes care of this possibility by treating each page
as a matrix before binding and unnesting their combination.
}
