% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error.R
\name{mm_error_body}
\alias{mm_error_body}
\title{Extract error body from a Megamation API response}
\usage{
mm_error_body(resp)
}
\arguments{
\item{resp}{An API response.}
}
\value{
A string from indexing the parsed response's \code{detail} list name.
}
\description{
Extract error body from a Megamation API response
}
