% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error.R
\name{mm_error_body}
\alias{mm_error_body}
\title{Extract error body from a Megamation API response}
\usage{
mm_error_body(resp)
}
\arguments{
\item{resp}{An API response.}
}
\value{
A string from the \code{detail} index of a named list (the parsed
response).
}
\description{
\code{mm_error_body()} simply plucks the \code{detail} list from the response body,
where Megamation's API includes informative error messages.
}
\examples{
fake_mm_resp <- httr2::response_json(
  body = list(
    detail = "This is a fake detail/message from the API's response."
  )
)
megamation:::mm_error_body(fake_mm_resp)
}
\keyword{internal}
