% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{mm_get}
\alias{mm_get}
\title{Perform a GET request to Megamation's API}
\usage{
mm_get(
  endpoint,
  ...,
  .get = "data",
  .url = get_env_url(),
  .key = get_env_key(),
  .full_resp = C(FALSE, TRUE)
)
}
\arguments{
\item{endpoint}{The API endpoint. For example,
\code{"timecard"} for employee transactions, and
`"workorder"`` for work orders. All endpoints are listed at
"https://apidocs.megamation.com/".}

\item{...}{<\link{dynamic-dots}> Name-value pairs to filter the request.
The name should be the lower-case name of a
field that is filter-enabled
(in Megamation's words, a criteria).

The value should be:
\itemize{
\item A length-1 atomic vector representing the value of said variable.
\item An as-is, modifier-value(s) combination,
e.g. \code{I("<>2020-01-01,2020-12-31")}. \code{\link[=I]{I()}} is used to prevent automatic
escaping.
}

To filter a single variable by multiple values, repeat name-value pairs with
the same name but different values. e.g. \verb{trade = "DM", trade = "PCO"}.}

\item{.get}{A length-1 character vector representing whether the request is
for the endpoint's \code{"data"}, \code{"criteria"}, \code{"labels"}, or \code{"schema"}.}

\item{.url}{The API URL.}

\item{.key}{The API key.}
}
\value{
A data frame of class \link{tbl} containing the requested information.
}
\description{
\code{mm_get()} accomplishes the full process of a GET request:
\itemize{
\item Creates an API request and defines its behaviour.
\item Performs the request and fetches the response.
\item Converts the body of the response to a data frame.
}

Where applicable, pagination is automatically applied to the request
by \code{\link[=mm_req_paginate]{mm_req_paginate()}} and returned pages are automatically combined.
}
