% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform.R
\name{mm_get}
\alias{mm_get}
\title{Get data}
\usage{
mm_get(endpoint, ..., .paginate = TRUE)
}
\arguments{
\item{endpoint}{The API endpoint. For example,
\code{"timecard"} for employee transactions, and \code{"workorder"}
for work orders. All endpoints are listed at
https://apidocs.megamation.com/.}

\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> Name-value pairs to filter the request.
The name should be the lower-case name of a
field that is filter-enabled
(in Megamation's words, a criteria).}

\item{.paginate}{If \code{TRUE}, paginate the request.}
}
\value{
A data frame of class \code{\link[tibble:tbl_df-class]{tbl_df}}
containing the requested information.
}
\description{
\code{mm_get()} accomplishes the full process of a GET request:
\itemize{
\item Creates an API request and defines its behaviour.
\item Performs the request and fetches the response.
\item Converts the body of the response to a data frame.
}

Where applicable, pagination is automatically applied to the request
by \code{\link[=mm_req_paginate]{mm_req_paginate()}} and returned pages are automatically combined.
}
\examples{
\dontrun{
# You can supply vectors to filtering variables
mm_get("workorder", wo_no = c("00001", "00002"))

# You can supply API modifiers when filtering
mm_get("workorder", trade = "[]PCO")

# You must supply date types to the date filter
jan_2023 <- seq.Date(
  as.Date("2023-01-01"),
  as.Date("2023-01-31"),
  by = "day"
  )

mm_get("employee", date = jan_2023)
}
}
