% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mm_get.R
\name{mm_get}
\alias{mm_get}
\title{Perform a GET request to Megamation's API}
\usage{
mm_get(
  endpoint,
  ...,
  .give = "data",
  .url = get_url(),
  .key = get_key(),
  .full_resp = C(FALSE, TRUE)
)
}
\arguments{
\item{endpoint}{The endpoint endpoint. For example,
"timecard" for employee transactions, and
"workorder" for work orders. All endpoints are listed here:
"https://apidocs.megamation.com/".}

\item{...}{API parameters to filter the request.}

\item{.give}{What the response should give:
\code{"data"}, \code{"criteria"}, \code{"labels"}, or \code{"schema"}.}

\item{.url}{The API URL.}

\item{.key}{The API key.}
}
\description{
Perform a GET request to Megamation's API
}
