% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{mm_keep_df}
\alias{mm_keep_df}
\title{Keep only data from parsed response}
\usage{
mm_keep_df(parsed)
}
\arguments{
\item{parsed}{Parsed response body.}
}
\value{
The \link{data.frame} object contained in the
parsed response.
}
\description{
\code{mm_keep_df()} extracts a data frame from a multi-layer list
(parsed body of an API response).
}
\keyword{internal}
