% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{mm_keep_df}
\alias{mm_keep_df}
\title{Keep only data from the JSON list returned by Megamation's API}
\usage{
mm_keep_df(list_from_json)
}
\arguments{
\item{list_from_json}{A list returned by \code{\link[jsonlite:fromJSON]{jsonlite::fromJSON()}}.}
}
\description{
Keep only data from the JSON list returned by Megamation's API
}
\keyword{internal}
