% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paginate.R
\name{mm_next_req}
\alias{mm_next_req}
\title{Return next request (if existing) from a Megamation API response}
\usage{
mm_next_req(resp, req)
}
\arguments{
\item{resp}{An API response.}

\item{req}{An API request generated by \code{\link[=mm_req]{mm_req()}}.}
}
\value{
An object of class \code{httr2_request} or \code{NULL}.
}
\description{
\code{mm_next_req()} returns the next request based on the \code{next_url} from
the parsed response.
}
\keyword{internal}
