% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mm_pull.R
\name{mm_pull}
\alias{mm_pull}
\title{Define and perform a Megamation API request}
\usage{
mm_pull(endpoint, ..., .for = "data", .url = get_env_url(), .key = get_env_key())
}
\arguments{
\item{endpoint}{The endpoint endpoint. For example,
"timecard" for employee transactions, and
"workorder" for work orders. All endpoints are listed here:
"https://apidocs.megamation.com/".}

\item{...}{API parameters to filter the request.}

\item{url}{The API URL.}

\item{key}{The API key.}
}
\description{
Define and perform a Megamation API request
}
