% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mm_req.R
\name{mm_req}
\alias{mm_req}
\title{Perform a Megamation API request}
\usage{
mm_req(url, resource, ..., username = "APIDL", api_key = get_api_key())
}
\arguments{
\item{url}{A URL starting with "https://api.megamation.com/".}

\item{resource}{The resource endpoint. For example,
"timecard" for employee transactions, and
"workorder" for work orders. All endpoints are listed here:
"https://apidocs.megamation.com/".}

\item{...}{API parameters to filter the request.}

\item{username}{The API key username.}

\item{api_key}{The API key.}
}
\description{
Perform a Megamation API request
}
