% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mm_req.R
\name{mm_req}
\alias{mm_req}
\title{Define a Megamation API request}
\usage{
mm_req(endpoint, ..., .for = "data", .url = get_env_url(), .key = get_env_key())
}
\arguments{
\item{endpoint}{The endpoint endpoint. For example,
"timecard" for employee transactions, and
"workorder" for work orders. All endpoints are listed here:
"https://apidocs.megamation.com/".}

\item{...}{Name-value pairs to filter the request.
The name should be the lower-case name of a
field that is filter-enabled
(that is, a criteria in Megamation's words).

The value should can be:
\itemize{
\item A character value of said variable.
\item An as-is, modifier-value(s) combination,
e.g. \code{I("<>2020-01-01,2020-12-31")}.
}

To filter a single variable by multiple values, repeat name-value pairs for
the same name but different values. e.g. \verb{trade = "DM", trade = "PCO"}.}

\item{.for}{What the response should give:
\code{"data"}, \code{"criteria"}, \code{"labels"}, or \code{"schema"}.}

\item{.url}{The API URL.}

\item{.key}{The API key.}
}
\description{
Define a Megamation API request
}
