% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mm_req.R
\name{mm_req}
\alias{mm_req}
\title{Define a Megamation API request}
\usage{
mm_req(resource, ..., url = get_url(), key = get_key())
}
\arguments{
\item{resource}{The resource endpoint. For example,
"timecard" for employee transactions, and
"workorder" for work orders. All endpoints are listed here:
"https://apidocs.megamation.com/".}

\item{...}{API parameters to filter the request.}

\item{url}{The API URL.}

\item{key}{The API key.}
}
\description{
Define a Megamation API request
}
