% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform.R
\name{mm_req_append}
\alias{mm_req_append}
\title{Append a GET request}
\usage{
mm_req_append(req, x)
}
\arguments{
\item{req}{An API request generated by \code{\link[=mm_req]{mm_req()}}.}

\item{x}{\code{"criteria"}, \code{"labels"}, or \code{"schema"}.}
}
\description{
Append a GET request
}
