% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mm_req_paginate.R
\name{mm_req_paginate}
\alias{mm_req_paginate}
\title{Add pagination to a Megamation API request}
\usage{
mm_req_paginate(req)
}
\arguments{
\item{req}{An API request generated by package httr2 or mm_req().}
}
\description{
Add pagination to a Megamation API request
}
