% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paginate.R
\name{mm_req_paginate}
\alias{mm_req_paginate}
\title{Add pagination to a Megamation API request}
\usage{
mm_req_paginate(req)
}
\arguments{
\item{req}{An API request generated by \code{\link[=mm_req]{mm_req()}}.}
}
\value{
An object of class \code{httr2_request}.
}
\description{
\code{mm_req_paginate()} adds pagination to a request based on the next_url given
by the parsed response.
}
\examples{
\dontshow{if (httr2::secret_has_key("HTTR2_KEY")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
mm_req("workorder") |> mm_req_paginate()
\dontshow{\}) # examplesIf}
}
