% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{mm_req_params}
\alias{mm_req_params}
\title{Modify request URL with filtering components}
\usage{
mm_req_params(req, ..., allfields = TRUE)
}
\arguments{
\item{req}{An API request generated by \code{\link[=mm_req]{mm_req()}}.}

\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> Name-value pairs to filter the request.
The name should be the lower-case name of a
field that is filter-enabled
(in Megamation's words, a criteria).}

\item{allfields}{If \code{TRUE}, return all fields currently available for
the endpoint.}
}
\value{
An object of class \code{httr2_request} with a pagination policy.
}
\description{
\code{mm_req_params()} adds filters to the request. By default, it adds the query
for all (currently available) fields.
}
\examples{
\dontshow{if (httr2::secret_has_key("HTTR2_KEY")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# No parameters
mm_req("status") |> mm_req_params()

# Multiple parameters
from <- as.Date("2023-09-20")
to <- as.Date("2023-10-20")
date <- seq(from, to, by = "day")
trade <- c("[]PCO", "[]DM")
mm_req("status") |> mm_req_params(date = date, trade = trade)
\dontshow{\}) # examplesIf}
}
