% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mm_req_perform.R
\name{mm_req_perform}
\alias{mm_req_perform}
\title{Perform a Megamation API request}
\usage{
mm_req_perform(req)
}
\arguments{
\item{req}{An API request generated by \code{\link[=mm_req]{mm_req()}} or \code{\link[httr2:req]{httr2::req()}}.}
}
\description{
Perform a Megamation API request
}
