% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform.R
\name{mm_req_perform}
\alias{mm_req_perform}
\title{Perform a Megamation API request}
\usage{
mm_req_perform(req, max_reqs = Inf)
}
\arguments{
\item{req}{An API request generated by \code{\link[=mm_req]{mm_req()}}.}

\item{max_reqs}{The maximum number of requests to perform. Use \code{Inf} to
perform all requests until \code{next_req()} returns \code{NULL}.}
}
\value{
A list of HTTP responses. Each response is an S3 list with class
\code{httr2_response}. (For more on this class, see \link[httr2:response]{httr2::response}.) If the
request was paginated, the responses correspond to pages.
}
\description{
After creating a request with \code{\link[=mm_req]{mm_req()}}, call \code{mm_req_perform()} to perform
it with pagination and fetch the results back to R.
}
\keyword{internal}
