% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform.R
\name{mm_req_perform}
\alias{mm_req_perform}
\title{Perform a Megamation API request}
\usage{
mm_req_perform(req)
}
\arguments{
\item{req}{An API request generated by \code{\link[=mm_req]{mm_req()}}.}
}
\value{
A list of HTTP responses. Each response is
an S3 list with class \code{httr2_response}. (For more on this class,
see \link[httr2:response]{httr2::response}.) If the request was paginated, these
responses correspond to pages.
}
\description{
After creating a request with \code{\link[=mm_req]{mm_req()}},
call \code{mm_req_perform()} to perform it and fetch
the results back to R.

Where the request is paginated, pagination is automatically performed
by \code{\link[httr2:req_perform_iteratively]{httr2::req_perform_iteratively()}}.
}
\examples{
\dontshow{if (httr2::secret_has_key("HTTR2_KEY")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
\dontshow{\}) # examplesIf}
}
