% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{mm_resp_extract}
\alias{mm_resp_extract}
\title{Extract data from Megamation API response}
\usage{
mm_resp_extract(resp)
}
\arguments{
\item{resp}{An API response.}
}
\value{
A data frame containing the endpoint data.
}
\description{
The body of the response contains raw bytes. After converting
these bytes to a string, encoding is done to resolve a UTF-8 issue from
Megamation's side.
}
\details{
\code{mm_resp_extract()} parses the raw bytes from an API response, and extracts
data from the parsed object.
}
\examples{
# Returns NULL from an empty response body
resp <- httr2::response_json()
resp |> megamation:::mm_resp_extract()
}
\keyword{internal}
