% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-parse.R
\name{mm_resp_parse}
\alias{mm_resp_parse}
\title{Parse body from Megamation API response}
\usage{
mm_resp_parse(resp)
}
\arguments{
\item{resp}{An API response.}
}
\value{
A list.
}
\description{
\code{mm_resp_parse()} parses the raw bytes from an API response, returning an R
list object. After converting bytes to characters, encoding is done to
resolve a UTF-8 issue from Megamation's side.
}
\examples{
# Parses empty response body to return an empty list
resp <- httr2::response_json()
resp |> megamation:::mm_resp_parse()
}
\keyword{internal}
