% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mm_set_creds.R
\name{mm_set_creds}
\alias{mm_set_creds}
\title{Install Megamation credentials in your \code{.Renviron} file for repeated use}
\usage{
mm_set_creds(key, url, overwrite = FALSE, install = FALSE, report = FALSE)
}
\arguments{
\item{key}{The API key provided to you by Megamation formatted in quotes.}

\item{url}{The API URL provided to you by Megamation
formatted in quotes.}

\item{overwrite}{If TRUE, will overwrite existing Megamation
credentials that you already have in your \code{.Renviron} file.}

\item{install}{If TRUE, will install the key in your \code{.Renviron} file
for use in future sessions.  Defaults to FALSE (single session use).}

\item{report}{If TRUE, ignores other arguments and outputs credentials as a
2-element named vector.}
}
\description{
This function adds your Megamation API key and base URL to your
\code{.Renviron} file so it can be called securely without being stored in
your code. After you have installed these two credentials, they can be
called any time with \code{Sys.getenv("MEGAMATION_KEY")} or
\code{Sys.getenv("MEGAMATION_URL")}. If you do not have an
\code{.Renviron} file, the function will create one for you. If you already
have an \code{.Renviron} file, the function will append the key to your
existing file, while making a backup of your original file for disaster
recovery purposes.
}
\examples{

\dontrun{
mm_set_creds(
  key = "<YOUR-MEGAMATION_KEY>",
  url = "<YOUR-MEGAMATION_URL>",
  install = TRUE
)
# Reload your environment so you can use the credentials without restarting R
readRenviron("~/.Renviron")
# You can check it with:
Sys.getenv("MEGAMATION_KEY")

# If you need to overwrite existing credentials:
mm_set_creds(
  key = "<YOUR-MEGAMATION_KEY>",
  url = "<YOUR-MEGAMATION_URL>",
  install = TRUE,
  overwrite = TRUE,
  install = TRUE
)
# Reload your environment to use the credentials
}
}
