% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creds.R
\name{mm_set_creds}
\alias{mm_set_creds}
\title{Install Megamation credentials in your \code{.Renviron} file for repeated use}
\usage{
mm_set_creds(key, url, overwrite = FALSE)
}
\arguments{
\item{key}{The API key provided to you by Megamation formatted in quotes.}

\item{url}{The API base URL provided to you by Megamation
formatted in quotes.}

\item{overwrite}{If TRUE, will overwrite existing Megamation
credentials that you already have in your \code{.Renviron} file.}
}
\description{
This function adds your Megamation API key and base URL to your
\link{.Renviron} file so it can be called securely without being stored in
your code. After you have installed these two credentials, \code{\link[=mm_request]{mm_request()}} and
\code{\link[=mm_get]{mm_get()}} will call them automatically. They can be
called manually at any time with \code{Sys.getenv("MEGAMATION_KEY")} or
\code{Sys.getenv("MEGAMATION_URL")}.

If you do not have a
\code{.Renviron} file, the function will create one for you. If you already
have a \code{.Renviron} file, the function will append the key to your
existing file, while making a backup of your original file for disaster
recovery purposes.
}
\examples{

\dontrun{
mm_set_creds(
  key = "<YOUR-MEGAMATION_KEY>",
  url = "<YOUR-MEGAMATION_URL>"
)
# Reload your environment so you can use the credentials
# without restarting R
readRenviron("~/.Renviron")
# You can check it with:
Sys.getenv("MEGAMATION_KEY")

# If you need to overwrite existing credentials:
mm_set_creds(
  key = "<YOUR-MEGAMATION_KEY>",
  url = "<YOUR-MEGAMATION_URL>",
  overwrite = TRUE
)
# Reload your environment to use the credentials
}
}
