% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parsed_to_tbl}
\alias{parsed_to_tbl}
\title{Create tibble from parsed response}
\usage{
parsed_to_tbl(parsed, .get = c("data", "criteria", "labels", "schema"))
}
\arguments{
\item{parsed}{Parsed response body.}

\item{.get}{A length-1 character vector representing whether the request is
for the endpoint's \code{"data"}, \code{"criteria"}, \code{"labels"}, or \code{"schema"}.}
}
\value{
A data frame of class \link{tbl_df} containing either
\itemize{
\item Columns of the endpoint's data.
\item Columns representing the list given by the endpoint's criteria, labels,
or schema: \code{list_name} and \code{value}.
}
}
\description{
\code{parsed_to_tbl()} constructs a data frame of class \link{tbl_df}.
The input should be a list returned from parsing
the API's response body.
}
