% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remove_api_urls}
\alias{remove_api_urls}
\title{Remove API \verb{_links} column in data}
\usage{
remove_api_urls(.data)
}
\arguments{
\item{.data}{Data within parsed response body.}
}
\value{
A data frame containing the endpoint data without a \verb{_links} column.
}
\description{
Remove API \verb{_links} column in data
}
\examples{
mtcars[, "_links"] <- "some-api-url.com"
mtcars |> remove_api_urls() |> names()
}
\keyword{internal}
